Improved git config installation
Instead of copying over the config file from a template there
is now an installer for git that reads one git config line at a
time and sets it if it is missing.
This way the config file is not overwritten when install is run
again.

Also improved code for reading in user name and email by using
readline.